Separate system tests into two files
